examples/flows/chat/chat-with-pdf-agent/requirements.txt
agent-framework-azure-ai
agent-framework-core
PyPDF2
faiss-cpu
openai
jinja2
python-dotenv
tiktoken
ipykernel
